+++ gulpfile.js
@@ -5,12 +5,28 @@ const mapappsBrowserSync = require("ct-mapapps-browser-sync");
 const isProduction = process.env.NODE_ENV === "production";
 console.info(`Configuring gulp build for ${isProduction ? "production" : "development"}`);
 
+const localOverrides = (function () {
+    if (isProduction) {
+        // Never override defaults in production mode
+        return undefined;
+    }
+
+    try {
+        return require("./gulpfile.overrides");
+    } catch (e) {
+        // File may not exist
+        return undefined;
+    }
+})();
+
 // used to transport test urls in "run-browser-tests-local" task
 const runBrowserTests = [];
 
 mapapps.registerTasks({
+    /** Enable debug logging */
+    debug: localOverrides?.debug ?? false,
     /** enable linting */
-    lintOnWatch: true,
+    lintOnWatch: localOverrides?.lintOnWatch ?? true,
     /** enable es6 by default */
     forceTranspile: true,
     /* A detailed description of available setting is available at https://www.npmjs.com/package/ct-mapapps-gulp-js */
@@ -18,6 +34,19 @@ mapapps.registerTasks({
 
     /* build source maps as e.g. ".js.map" */
     sourceMaps: "file",
+
+    /** Build Unit-Tests only in dev mode */
+    rollupBuildTests: !isProduction,
+    /** Amount of Threads used to build the bundles, if there are only a few bundles 1 is ok.
+     *  More as 3 is normally not required.
+     */
+    rollupBuildMaxWorkers: localOverrides?.rollupBuildMaxWorkers ?? 1,
+
+    /** List of build time flags, usage like: import { debug } from "build-config!".
+     */
+    rollupConfig: {
+        debug: !isProduction
+    },
 
     /* a list of themes inside this project */
     themes: ['theme-custom'],
@@ -39,15 +68,21 @@ mapapps.registerTasks({
     },
     runBrowserTests,
     watchFinishedReceiver() {
-        mapappsBrowserSync.state.reload();
+        if (localOverrides?.autoReload ?? true) {
+            mapappsBrowserSync.state.reload();
+        }
     }
 }, gulp);
 
 mapappsBrowserSync.registerTask({
-    port: 9090,
+    // on which port to listen
+    port: localOverrides?.port ?? 9090,
     // activate https protocol, generates a self signed certificate for "localhost"
     // https://browsersync.io/docs/options#option-https
-    https: false,
+    https: localOverrides?.https ?? false,
+
+    // to prevent auto open of browser, set this to false
+    urlToOpen: localOverrides?.openBrowser ?? true,
 
     jsreg: {
         //npmDir : __dirname + "/node_modules/",
@@ -58,9 +93,7 @@ mapappsBrowserSync.registerTask({
         ]
     },
     // prevent reload by browser sync (reload triggered on watch end)
-    externalReloadTrigger: true,
-    // to prevent auto open of browser, set this to false
-    urlToOpen: true
+    externalReloadTrigger: true
 }, gulp);
 
 gulp.task("build",
@@ -69,6 +102,7 @@ gulp.task("build",
         "themes-copy",
         gulp.parallel(
             "js-transpile",
+            "rollup-build",
             "themes-compile"
         )
     )