+++ src/main/js/bundles/sample_camera/module.js
@@ -1,1 +1,1 @@
-import "./CameraControlFactory";
+export { default as CameraControlFactory } from "./CameraControlFactory";