+++ src/main/js/bundles/sample_camera/tests/all.js
@@ -1,1 +1,1 @@
-import "./CameraControls";
+import "./CameraControlsTest";