+++ src/main/js/bundles/sample_camera/tests/intern-all.js
@@ -1,1 +1,1 @@
-import "./CameraControls_OLD_INTERN";
+import "./CameraControlsTest_OLD_INTERN";